+++ CHANGELOG.md
@@ -52,6 +52,9 @@ four, three, and two argument variations.
 `(err, ctx, state)`, rather than `(err, ok, state)`, where `ctx` is an object,
 rather than a boolean, and contains the protocol context needed to validate the
 authentication response.
+- Switched to using `OAuth2#get`, from `OAuth2#_request`, when making UserInfo
+request.  As a result, `Accept: 'application/json` header no longer sent.  This
+header isn't needed, per spec.
 
 ### Removed